cleanup: Complete lib/tools architectural boundary violations
🧠 Sequential Thinking: Comprehensive Analysis Results

### 🚨 CRITICAL FINDINGS: Initial Analysis Was Incomplete

❌ **Previous Analysis**: Only identified 2 exact duplicates
✅ **Complete Analysis**: Identified systematic architectural boundary violations

### 🏗️ **Comprehensive Architectural Cleanup:**

#### **1. Empty Directory Cleanup:**
- ❌ **REMOVED**: tools/security/ (empty directory)
- ❌ **REMOVED**: lib/core/utils/ (empty directory)

#### **2. Setup Architecture Corrected:**
- ✅ **MOVED**: tools/automation/setup_stakeholder_management.py → tools/setup/
- ✅ **MOVED**: tools/automation/setup_task_tracking.py → tools/setup/
- ✅ **CREATED**: tools/setup/ directory (required by P0 tests)

#### **3. Security Architecture Consolidated:**
- ✅ **MOVED**: tools/quality/stakeholder_name_scanner.py → lib/security/scanners/
- ✅ **MOVED**: tools/quality/sensitive_data_scanner.py → lib/security/scanners/
- ✅ **UPDATED**: lib/security/scanners/__init__.py for proper exports
- ✅ **FIXED**: .pre-commit-config.yaml to reference new scanner location

#### **4. Exact Duplicates Eliminated:**
- ❌ **REMOVED**: tools/automation/setup_meeting_intelligence.py (314 lines)
- ❌ **REMOVED**: tools/automation/setup_smart_git.py (373 lines)
- ✅ **CANONICAL**: lib/setup/ versions maintained

### 📊 **Clear Architectural Boundaries Established:**

#### **lib/** = Core System Functionality & Importable Modules
- ✅ **lib/utils/**: Importable utilities (formatting.py, parallel.py)
- ✅ **lib/automation/**: Core automation logic (task_manager.py)
- ✅ **lib/security/**: Security functionality (scanners/, security_manager.py)
- ✅ **lib/setup/**: Core setup modules (git_automation.py, meeting_intelligence.py)

#### **tools/** = Development & Operational Utilities
- ✅ **tools/setup/**: Setup scripts and operational utilities
- ✅ **tools/automation/**: 18+ operational scripts and daily alerts
- ✅ **tools/quality/**: Development quality checks and validators
- ✅ **tools/ci/**: CI/CD pipeline utilities
- ✅ **tools/validator/**: Code validation utilities

### 🎯 **P0 Test Compliance:**
- ✅ **Setup Feature P0**: PASSING (tools/setup/ directory restored)
- ✅ **Performance P0**: PASSING (resource efficiency maintained)
- ✅ **All 40 P0 tests**: Ready for validation

### 🚀 **Result:**
Complete architectural clarity with zero boundary confusion.
Core functionality properly separated from operational utilities.
