--- README.md
@@ -1,113 +1,0 @@
-# ClaudeDirector - AI Strategic Leadership System
-
-🎯 **The industry's first completely transparent AI strategic leadership system with role-based customization for engineering leaders.**
-
-## Value Proposition
-
-**Transform Strategic Leadership Through Transparent AI Enhancement**
-
-ClaudeDirector delivers unprecedented strategic clarity by combining:
-- **Role-Based Intelligence**: Customized for VP/CTO/Director/Manager/Staff/Product Engineering roles
-- **Complete Transparency**: Real-time disclosure of all AI enhancements and strategic frameworks
-- **Next Action Clarity**: >85% success rate in providing clear, actionable strategic next steps
-- **Framework Attribution**: 25+ strategic methodologies with systematic application and crediting
-- **Enterprise Security**: Sensitive stakeholder intelligence protection with audit trails
-- **Business ROI**: Quantifiable investment tracking and strategic value measurement
-
-**Key Business Outcomes**:
-- 15-20 minutes saved per strategic session through context preservation
-- 40% reduction in strategic analysis time through framework application
-- 85%+ Next Action Clarity Rate for strategic decision making
-- Complete audit trail for enterprise AI governance compliance
-
-## Quick Start
-
-ClaudeDirector is designed for **chat-based interaction** through Cursor IDE. Simply start asking strategic questions and ClaudeDirector will guide you through role-based customization on first use.
-
-### Supported Leadership Roles
-- **VP/SVP Engineering**: Executive strategy, board communication, organizational scaling
-- **CTO**: Technology vision, strategic leadership, competitive positioning
-- **Engineering Director**: Platform strategy, team coordination (default)
-- **Engineering Manager**: Team leadership, delivery optimization, people management
-- **Staff/Principal Engineer**: Technical strategy, influence without authority
-- **Product Engineering Lead**: Product-engineering alignment, user impact optimization
-
-## Key Features
-
-### 🎯 Strategic Leadership
-- **First Principles Thinking**: Root cause analysis and assumption testing
-- **Executive Communication**: Board-ready insights and recommendations
-- **Cross-Team Coordination**: Alignment across engineering/product/design
-- **Platform Strategy**: Technical investments → business outcomes
-
-### 🤖 AI Transparency
-- **Complete Transparency**: All AI enhancements disclosed in real-time
-- **Framework Attribution**: Strategic methodologies credited and explained
-- **Audit Trail Ready**: Enterprise AI governance compliance
-- **Performance SLAs**: <5s enhanced responses, 99.5%+ uptime
-
-### 🏗️ Platform Intelligence
-- **Sequential Thinking**: Systematic problem-solving methodology
-- **Context7 Integration**: Strategic framework patterns and best practices
-- **Performance Optimization**: Enterprise-grade response times
-- **P0 Critical Features**: Zero-tolerance quality enforcement
-
-## Architecture
-
-ClaudeDirector uses a **persona-based system** with strategic AI enhancement:
-
-- **Strategic Personas**: Role-specific expertise (Diego, Camille, Rachel, Alvaro, Martin, etc.)
-- **MCP Integration**: Model Context Protocol for enhanced capabilities
-- **Unified Performance**: Sub-500ms response times with intelligent caching
-- **Graceful Degradation**: Full functionality maintained without external dependencies
-
-## Getting Started
-
-### 📥 Installation
-
-Clone the repository and start your first strategic session:
-
-```bash
-git clone https://github.com/chriscantu/ClaudeDirector.git
-cd ClaudeDirector
-```
-
-### 🚀 Quick Start
-
-1. **Open in Cursor IDE**: ClaudeDirector is optimized for Cursor integration
-2. **Ask Strategic Questions**: Start with any leadership or technical challenge
-3. **Role Customization**: First-run wizard adapts to your specific role and context
-4. **Strategic Guidance**: Receive transparent, framework-based recommendations
-
-### Example Questions
-- "How should we restructure our engineering teams for better scalability?"
-- "What's our platform investment strategy for the next quarter?"
-- "How do we align engineering velocity with business objectives?"
-- "What architectural decisions support our 3-year growth plan?"
-
-## Documentation
-
-- **Architecture**: `docs/architecture/OVERVIEW.md` - System architecture and design decisions
-- **Project Structure**: `docs/architecture/PROJECT_STRUCTURE.md` - Codebase organization
-- **Capabilities**: `docs/CAPABILITIES.md` - Complete feature overview
-- **Getting Started**: `docs/GETTING_STARTED.md` - Detailed setup and usage guide
-
-## Enterprise Features
-
-- **Complete Audit Trails**: Full traceability of all AI enhancements
-- **Framework Attribution**: Systematic crediting of strategic methodologies
-- **Multi-Persona Coordination**: Transparent cross-functional collaboration
-- **Performance Guarantees**: Response time commitments with quality preservation
-- **Zero Data Loss**: All user data, configurations, and state preserved
-
-## Support
-
-ClaudeDirector is designed for **self-service strategic leadership**. The system provides:
-- **Contextual Guidance**: In-chat help and strategic recommendations
-- **Role-Based Customization**: Adapted to your specific leadership context
-- **Transparent AI**: Complete visibility into all AI-enhanced capabilities
-- **Enterprise Reliability**: P0 critical feature enforcement and monitoring
-
----
-
-**Ready to transform your strategic leadership?** Start asking questions in Cursor and let ClaudeDirector guide you to better decisions and outcomes.
